requirements.txt
discord.py
python-dotenv
flask
firebase-admin